requirements.txt
dnspython==2.2.1
pymongo==4.1.1
python-dotenv==0.20.0
adafruit-io==2.7.0
